remove unneeded unboxing

+++ src/main/java/com/bovilexics/javaph/JavaPH.java
@@ -550,7 +550,7 @@ public class JavaPH extends JApplet implements IconProvider {
 					@NotNull final String portText = portInt.toString();
 					portStatusLabel.setText(PORT_LABEL_PREFIX + portText + PORT_LABEL_SUFFIX);
 					
-					connection = new QiConnection(serverText, portInt.intValue());
+					connection = new QiConnection(serverText, portInt);
 
 					if (getLoadFields() == LOAD_FIELDS_SELECTED && !(server.getFieldState() == QiFieldState.FIELD_LOAD_ERROR) && !(server.getFieldState() == QiFieldState.FIELD_LOAD_TRUE))
 					{
